Inherit from templates/NWButton
