+++ web/components/input/buttons/PrimaryButton/index.js
@@ -1,22 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const PrimaryButton = ({ text, onClick, disabled }) => (
-  <span>
-    <button
-      onClick={onClick}
-      type="button"
-      disabled={disabled}
-      className="primary">
-      { text }
-    </button>
-  </span>
-);
+import NWButton from '../templates/NWButton';
 
-PrimaryButton.propTypes = {
+const SecondaryButton = props => <NWButton type="primary" {...props} />;
+
+SecondaryButton.propTypes = {
   text: PropTypes.string.isRequired,
   onClick: PropTypes.func,
   disabled: PropTypes.bool,
 };
 
-export default PrimaryButton;
+export default SecondaryButton;
